pyproject.toml
[project]
name = "final-team-9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=4.2.2",
    "langsmith>=0.7.34",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
